Taking Damage

+++ Rainbow Seven/Assets/Scripts/Interfaces/IDamageable.cs
@@ -1,4 +1,6 @@
+using Photon.Pun;
+
 public interface IDamageable 
 {
-    void TakeDamage(float damage);
+    void TakeDamage(int damage);
 }
